develop = main

+++ AuctionPlanningServiceAPI/Service/MongoDBService.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Security.Cryptography;
 using AuctionPlanningServiceAPI.Controllers;
 using AuctionPlanningServiceAPI.Model;
 using Microsoft.AspNetCore.Mvc;
@@ -14,88 +15,67 @@ namespace AuctionPlanningServiceAPI.Service
     public class MongoDBService : IAuctionPlanningRepository
 
     {
-        private readonly ILogger<AuctionPlanningServiceController> _logger;
-
+        private readonly ILogger<AuctionPlanningController> _logger;
         private readonly IConfiguration _config;
 
         private readonly string _connectionURI;
 
         private readonly string _auctionsDatabase;
-
         private readonly string _inventoryDatabase;
 
         private readonly string _auctionCollectionName;
-
         private readonly string _articleCollectionName;
+
 
         private readonly IMongoCollection<Auction> _auctionCollection;
         private readonly IMongoCollection<Article> _articleCollection;
 
-        public MongoDBService(ILogger<AuctionPlanningServiceController> logger, IConfiguration config)
+        public MongoDBService(ILogger<AuctionPlanningController> logger, IConfiguration config, EnviromentVariables vaultSecrets)
         {
             _logger = logger;
             _config = config;
 
             try
             {
-        
-            // // Auction database and collections
-            // _auctionsDatabase = config["AuctionsDatabase"] ?? "Auctionsdatabase missing";
-            // _auctionCollectionName = config["AuctionCollection"] ?? "Auctioncollection name missing";
+                // Retrieves enviroment variables from program.cs, from injected EnviromentVariables class
+                _connectionURI = vaultSecrets.dictionary["ConnectionURI"];
 
+                // Retrieves Auction database and collections
+                _auctionsDatabase = config["AuctionsDatabase"] ?? "Auctionsdatabase missing";
+                _auctionCollectionName = config["AuctionCollection"] ?? "Auctioncollection name missing";
 
-            // // Inventory database and collection
-            // _inventoryDatabase = config["InventoryDatabase"] ?? "Inventorydatabase missing";
-            // _articleCollectionName = config["ArticleCollection"] ?? "Articlecollection name missing";
+                // Retrieves Inventory database and collection
+                _inventoryDatabase = config["InventoryDatabase"] ?? "Inventorydatabase missing";
+                _articleCollectionName = config["ArticleCollection"] ?? "Articlecollection name missing";
 
-            // _logger.LogInformation($"AuctionService secrets: ConnectionURI: {_connectionURI}");
-            // _logger.LogInformation($"AuctionService Database and Collections: Auctiondatabase: {_auctionsDatabase}, Auctionsdatabase: {_auctionsDatabase}");
+                _logger.LogInformation($"AuctionService secrets: ConnectionURI: {_connectionURI}");
+                _logger.LogInformation($"AuctionService Database and Collections: Auctiondatabase: {_auctionsDatabase}, Auctionsdatabase: {_auctionCollectionName}");
+                _logger.LogInformation($"Inventory Database and Collections: Inventorydatabase: {_auctionsDatabase}, Articlecollection: {_articleCollectionName}");
 
             }
             catch (Exception ex)
             {
-                _logger.LogError("Error retrieving environment variables");
+                _logger.LogError("Error retrieving enviroment variables");
 
                 throw;
             }
-
-            _connectionURI = "mongodb://admin:1234@localhost:27018/";
-
-            //Auction database and collections
-            _auctionsDatabase = "Auctions";
-            _auctionCollectionName = "listings";
-
-            //Inventory database and collection
-            _inventoryDatabase = "inventoryDatabase";
-            _articleCollectionName = "article";
 
             try
             {
                 // Sets MongoDB client
                 var mongoClient = new MongoClient(_connectionURI);
-                _logger.LogInformation($"[*] CONNECTION_URI: {_connectionURI}");
-
 
                 // Sets MongoDB Database
                 var auctionsDatabase = mongoClient.GetDatabase(_auctionsDatabase);
-                _logger.LogInformation($"[*] DATABASE: {_auctionsDatabase}");
-
                 var inventoryDatabase = mongoClient.GetDatabase(_inventoryDatabase);
-                _logger.LogInformation($"[*] DATABASE: {_inventoryDatabase}");
-
 
                 // Collections
                 _auctionCollection = auctionsDatabase.GetCollection<Auction>(_auctionCollectionName);
-                _logger.LogInformation($"[*] COLLECTION: {_auctionCollectionName}");
-
                 _articleCollection = inventoryDatabase.GetCollection<Article>(_articleCollectionName);
-                _logger.LogInformation($"[*] COLLECTION: {_articleCollectionName}");
-
             }
             catch (Exception ex)
             {
                 _logger.LogError($"Error trying to connect to database: {ex.Message}");
-
                 throw;
             }
         }
@@ -105,26 +85,38 @@ namespace AuctionPlanningServiceAPI.Service
         {
             try
             {
-                _logger.LogInformation($"[*]: AddAuction(AuctionDTO auctionDTO) called: Adding a new auction instance to the database.\nHighestBid: {auctionDTO.HighestBid}\nBidCounter: {auctionDTO.BidCounter}\nStartDate: {auctionDTO.StartDate}\nEndDate: {auctionDTO.EndDate}\nViews: {auctionDTO.Views}\nArticleID: {auctionDTO.ArticleID}");
+                _logger.LogInformation($"POST: addAuction called, StartDate: {auctionDTO.StartDate}, EndDate: {auctionDTO.EndDate}, ArticleID: {auctionDTO.ArticleID}");
 
-                Auction auction = new Auction
+                Article auctionArticle = new Article();
+
+                auctionArticle = await _articleCollection.Find(x => x.ArticleID == auctionDTO.ArticleID).FirstOrDefaultAsync();
+
+                if (auctionArticle == null)
                 {
-                    AuctionID = ObjectId.GenerateNewId().ToString(),
-                    HighestBid = auctionDTO.HighestBid,
-                    BidCounter = auctionDTO.BidCounter,
-                    StartDate = auctionDTO.StartDate,
-                    EndDate = auctionDTO.EndDate,
-                    Views = auctionDTO.Views,
-                    ArticleID = auctionDTO.ArticleID,
-                };
+                    _logger.LogInformation("Error finding article for auction");
+                    return null;
+                }
+                else
+                {
+                    Auction auction = new Auction
+                    {
+                        AuctionID = ObjectId.GenerateNewId().ToString(),
+                        HighestBid = 0,
+                        Bids = new List<Bid>(),
+                        StartDate = auctionDTO.StartDate,
+                        EndDate = auctionDTO.EndDate,
+                        Views = 0,
+                        Article = auctionArticle,
+                    };
 
-                await _auctionCollection.InsertOneAsync(auction);
+                    await _auctionCollection.InsertOneAsync(auction);
 
-                return auction;
+                    return auction;
+                }
             }
             catch (Exception ex)
             {
-                _logger.LogError($"EXCEPTION CAUGHT: {ex.Message}");
+                _logger.LogError($"Error in add auction: {ex.Message}");
 
                 throw;
             }
@@ -132,7 +124,7 @@ namespace AuctionPlanningServiceAPI.Service
         //GET - Return a list of all auctions
         public async Task<List<Auction>> GetAllAuctions()
         {
-            _logger.LogInformation($"[*] GettAllAuctions() called: Fetching all auctions in the database");
+            _logger.LogInformation($"getAll endpoint called");
 
             try
             {
@@ -140,39 +132,52 @@ namespace AuctionPlanningServiceAPI.Service
 
                 allAuctions = await _auctionCollection.Find(_ => true).ToListAsync<Auction>();
 
+                if (allAuctions == null)
+                {
+                    _logger.LogInformation("No auctions found");
+                }
+
                 return allAuctions;
             }
             catch (Exception ex)
             {
-                _logger.LogError($"EXCEPTION CAUGHT: {ex.Message}");
+                _logger.LogError($"Failed calling getAllAuctions: {ex.Message}");
 
                 throw;
             }
 
         }
         //DELETE - Removes an auction
-        public async Task<Auction> DeleteAuction(string auctionId)
+        public async Task<Auction> DeleteAuction(string id)
         {
-            _logger.LogInformation($"[*] DeleteAuction(string auctionId) called: Deleting auction in the database with auctionId {auctionId}");
-
             try
             {
+                _logger.LogInformation($"DeleteAuction called with Auction ID = {id}");
+
                 Auction deleteAuction = new Auction();
 
-                deleteAuction = await _auctionCollection.Find(x => x.AuctionID == auctionId).FirstAsync<Auction>();
+                deleteAuction = await _auctionCollection.Find(x => x.AuctionID == id).FirstOrDefaultAsync();
 
-                FilterDefinition<Auction> filter = Builders<Auction>.Filter.Eq("AuctionID", auctionId);
+                if (deleteAuction == null)
+                {
+                    _logger.LogInformation("No auctions found to be deleted");
 
-                await _auctionCollection.DeleteOneAsync(filter);
+                    return null;
+                }
+                else
+                {
+                    FilterDefinition<Auction> filter = Builders<Auction>.Filter.Eq("AuctionID", id);
 
-                _logger.LogInformation($"Auction with id {auctionId} got deleted");
+                    await _auctionCollection.DeleteOneAsync(filter);
 
-                return deleteAuction;
+                    _logger.LogInformation($"id got deleted: {id}");
 
+                    return deleteAuction;
+                }
             }
             catch (Exception ex)
             {
-                _logger.LogError($"EXCEPTION CAUGHT: {ex.Message}");
+                _logger.LogError($"Failed while trying to deleteAuction: {ex.Message}");
 
                 throw;
             }
@@ -180,25 +185,26 @@ namespace AuctionPlanningServiceAPI.Service
         }
 
         // GET - Retrieves an auction by ID
-        public async Task<Auction> GetAuctionByID(string auctionId)
+        public async Task<Auction> GetAuctionByID(string id)
         {
-            _logger.LogInformation($"[*] GetAuctionByID(string auctionId) called: Fetching auction information from auctionId {auctionId}");
-
             try
             {
-                Auction auction = await _auctionCollection.Find(x => x.AuctionID == auctionId).FirstOrDefaultAsync();
+                _logger.LogInformation($"GET auction called with ID: {id}");
+
+                Auction auction = await _auctionCollection.Find(x => x.AuctionID == id).FirstOrDefaultAsync();
 
                 if (auction == null)
                 {
-                    _logger.LogError($"Error finding auction: {auctionId}");
+                    _logger.LogError($"Error finding auction: {id}");
+
+                    return null;
                 }
 
                 return auction;
             }
             catch (Exception ex)
             {
-                _logger.LogError($"EXCEPTION CAUGHT: {ex.Message}");
-
+                _logger.LogError($"Error in getAuction: {ex.Message}");
                 throw;
             }
         }
@@ -206,36 +212,43 @@ namespace AuctionPlanningServiceAPI.Service
 
 
         // PUT - Updates an auction
-        public async Task<Auction> UpdateAuction(string auctionId, AuctionDTO auctionDTO)
+        public async Task<Auction> UpdateAuction(string id, AuctionDTO auctionDTO)
         {
-            _logger.LogInformation($"[*] UpdateAuction(string auctionId, AuctionDTO auctionDTO) called: Updating the auction with auctiondId {auctionId}");
-
             try
             {
-                Auction existingAuction = await _auctionCollection.Find(x => x.AuctionID == auctionId).FirstOrDefaultAsync();
+                _logger.LogInformation($"PUT auction called with ID: {id}");
+
+                Auction existingAuction = await _auctionCollection.Find(x => x.AuctionID == id).FirstOrDefaultAsync();
 
                 if (existingAuction == null)
                 {
-                    _logger.LogError($"Error finding auction: {auctionId}");
+                    _logger.LogError($"Error finding auction: {id}");
+
+                    return null;
                 }
 
+                Article article = await _articleCollection.Find(x => x.ArticleID == auctionDTO.ArticleID).FirstOrDefaultAsync();
+
+                if (article == null)
+                {
+                    _logger.LogError($"Error finding article with ID: {auctionDTO.ArticleID}");
+
+                    return null;
+                }
                 // Update the properties of the existing auction with the values from the DTO
-                existingAuction.HighestBid = auctionDTO.HighestBid;
-                existingAuction.BidCounter = auctionDTO.BidCounter;
                 existingAuction.StartDate = auctionDTO.StartDate;
                 existingAuction.EndDate = auctionDTO.EndDate;
-                existingAuction.Views = auctionDTO.Views;
-                existingAuction.ArticleID = auctionDTO.ArticleID;
+                existingAuction.Article = article;
 
                 // Replace the existing auction in the database with the updated auction
-                await _auctionCollection.ReplaceOneAsync(x => x.AuctionID == auctionId, existingAuction);
+                await _auctionCollection.ReplaceOneAsync(x => x.AuctionID == id, existingAuction);
 
                 return existingAuction;
             }
             catch (Exception ex)
             {
-                _logger.LogError($"EXCEPTION CAUGHT: {ex.Message}");
-                
+                _logger.LogError($"Error in updateAuction: {ex.Message}");
+
                 throw;
             }
         }
